Bug Fixes and Some Features

+++ src/scratchpad.cpp
@@ -1,11 +1,11 @@
 #include "scratchpad.h"
 
-Scratchpad::Scratchpad(std::size_t capacity_bytes) : capacity_bytes_(capacity_bytes) {}
+Scratchpad::Scratchpad(std::uint64_t capacity_bytes) : capacity_bytes_(capacity_bytes) {}
 
-bool Scratchpad::can_hold(std::size_t bytes) const {
+bool Scratchpad::can_hold(std::uint64_t bytes) const {
     return bytes <= capacity_bytes_;
 }
 
-std::size_t Scratchpad::capacity_bytes() const {
+std::uint64_t Scratchpad::capacity_bytes() const {
     return capacity_bytes_;
 }
